Made some more progress -- see extended description
The instruction cache seems to be working. I ran a test program that was essentially a for loop and it was getting cache hits a fair amount of time. Guess that means that the big issues left are getting the hash right (right now i'm just pulling the bottom bits) and making sure the data cache works correctly. Hopefully Diet'z test plan comes through and helps us out.