Data cache and some prefetch --See extended
Added in functionality for the data cache and changed some minor errors about load and store that my group had made during the last project. Not sure if what i put for prefetch is totally correct, but in theory it makes sense, if we get a hit on the instruction cache then we might not be reading. Might have to change it so that it prefetches any time there isn't another read going (not just when we get a cache hit).